Only waits for background processing to finish if it's a front-end request, required caching the check for a front-end request which required moving the check to it's own class.
